tests/wpt/tests/css/css-backgrounds/parsing/background-repeat-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedStyle().backgroundRepeat</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-repeat">
<meta name="assert" content="background-repeat value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-repeat", "repeat-x");
test_computed_value("background-repeat", "repeat-y");
test_computed_value("background-repeat", "repeat");
test_computed_value("background-repeat", "space");
test_computed_value("background-repeat", "round");
test_computed_value("background-repeat", "no-repeat");
test_computed_value("background-repeat", "repeat space");
test_computed_value("background-repeat", "round no-repeat");
test_computed_value("background-repeat", "repeat repeat", "repeat");
test_computed_value("background-repeat", "repeat-x, repeat-y, repeat");

document.getElementById("target").style['background-image'] = "none, none";
test_computed_value("background-repeat", "repeat-x", "repeat-x", "multiple base values");
test_computed_value("background-repeat", "repeat-x, repeat-y", "repeat-x, repeat-y", "multiple base values");
test_computed_value("background-repeat", "repeat-x, repeat-y, repeat", "repeat-x, repeat-y, repeat", "multiple base values");
</script>
</body>
</html>
